test: merge tests and mocked data
